refactor(kernel): split plugin openclaw compatibility seam
