+++ crates/kernel/src/plugin.rs
@@ -15,6 +15,8 @@ use crate::{
     pack::VerticalPackManifest,
 };
 
+mod openclaw;
+
 pub const PACKAGE_MANIFEST_FILE_NAME: &str = "loong.plugin.json";
 const OPENCLAW_PACKAGE_MANIFEST_FILE_NAME: &str = "openclaw.plugin.json";
 const PACKAGE_JSON_FILE_NAME: &str = "package.json";
@@ -783,83 +785,6 @@ impl PackageManifestDocument {
     }
 }
 
-#[derive(Debug, Deserialize)]
-struct OpenClawManifestDocument {
-    id: String,
-    #[serde(default, rename = "configSchema")]
-    config_schema: Option<Value>,
-    #[serde(default, rename = "enabledByDefault")]
-    enabled_by_default: bool,
-    #[serde(default)]
-    kind: Option<String>,
-    #[serde(default)]
-    channels: Vec<String>,
-    #[serde(default)]
-    providers: Vec<String>,
-    #[serde(default, rename = "providerAuthEnvVars")]
-    provider_auth_env_vars: BTreeMap<String, Vec<String>>,
-    #[serde(default, rename = "providerAuthChoices")]
-    provider_auth_choices: Vec<Value>,
-    #[serde(default)]
-    skills: Vec<String>,
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    description: Option<String>,
-    #[serde(default)]
-    version: Option<String>,
-    #[serde(default, rename = "uiHints")]
-    ui_hints: BTreeMap<String, Value>,
-}
-
-#[derive(Debug, Deserialize, Default)]
-struct OpenClawPackageJsonDocument {
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    version: Option<String>,
-    #[serde(default)]
-    description: Option<String>,
-    #[serde(default)]
-    openclaw: Option<OpenClawPackageMetadataDocument>,
-}
-
-#[derive(Debug, Deserialize, Default)]
-struct OpenClawPackageMetadataDocument {
-    #[serde(default)]
-    extensions: Vec<String>,
-    #[serde(default, rename = "setupEntry")]
-    setup_entry: Option<String>,
-    #[serde(default)]
-    channel: Option<OpenClawPackageChannelDocument>,
-    #[serde(default)]
-    install: Option<OpenClawPackageInstallDocument>,
-}
-
-#[derive(Debug, Deserialize, Default)]
-struct OpenClawPackageChannelDocument {
-    #[serde(default)]
-    id: Option<String>,
-    #[serde(default)]
-    label: Option<String>,
-    #[serde(default, rename = "docsPath")]
-    docs_path: Option<String>,
-    #[serde(default)]
-    blurb: Option<String>,
-    #[serde(default)]
-    aliases: Vec<String>,
-}
-
-#[derive(Debug, Deserialize, Default)]
-struct OpenClawPackageInstallDocument {
-    #[serde(default, rename = "npmSpec")]
-    npm_spec: Option<String>,
-    #[serde(default, rename = "localPath")]
-    local_path: Option<String>,
-    #[serde(default, rename = "minHostVersion")]
-    min_host_version: Option<String>,
-}
-
 fn collect_files(path: &Path, acc: &mut Vec<PathBuf>) -> Result<(), IntegrationError> {
     let metadata = fs::metadata(path).map_err(|error| IntegrationError::PluginFileRead {
         path: path.display().to_string(),
@@ -906,13 +831,15 @@ fn collect_package_manifest_descriptors(
         }
 
         if is_openclaw_package_manifest_file(path) {
-            let descriptor = parse_openclaw_manifest_descriptor(path)?;
+            let descriptor = openclaw::parse_openclaw_manifest_descriptor(path)?;
             descriptors.insert(PathBuf::from(descriptor.path.clone()), descriptor);
             continue;
         }
 
         if is_package_json_file(path) {
-            for descriptor in parse_openclaw_legacy_package_descriptors(path, &known_files)? {
+            for descriptor in
+                openclaw::parse_openclaw_legacy_package_descriptors(path, &known_files)?
+            {
                 descriptors.insert(PathBuf::from(descriptor.path.clone()), descriptor);
             }
         }
@@ -1130,584 +1057,6 @@ fn parse_package_manifest_file(path: &Path) -> Result<PluginManifest, Integratio
     )?;
 
     Ok(normalized_manifest)
-}
-
-fn parse_openclaw_manifest_descriptor(path: &Path) -> Result<PluginDescriptor, IntegrationError> {
-    let document = parse_json_document::<OpenClawManifestDocument>(path)?;
-    validate_openclaw_manifest_document(&document, path)?;
-
-    let package_json_path = path
-        .parent()
-        .map(|parent| parent.join(PACKAGE_JSON_FILE_NAME))
-        .filter(|candidate| candidate.is_file());
-    let package_document = package_json_path
-        .as_deref()
-        .map(parse_json_document::<OpenClawPackageJsonDocument>)
-        .transpose()?;
-
-    let package_root = path.parent().unwrap_or(path);
-    let primary_entry_path =
-        resolve_openclaw_primary_entry_path(package_root, package_document.as_ref(), true);
-    let setup_entry_path = package_document
-        .as_ref()
-        .and_then(|package| package.openclaw.as_ref())
-        .and_then(|metadata| metadata.setup_entry.as_deref())
-        .and_then(|entry| resolve_openclaw_relative_path(package_root, entry));
-    let manifest = build_openclaw_manifest(
-        &document,
-        package_document.as_ref(),
-        primary_entry_path.as_deref(),
-        setup_entry_path.as_deref(),
-        PluginCompatibilityMode::OpenClawModern,
-    );
-    let descriptor_path = primary_entry_path.as_deref().unwrap_or(path);
-    let descriptor = build_plugin_descriptor(
-        descriptor_path,
-        PluginSourceKind::PackageManifest,
-        PluginContractDialect::OpenClawModernManifest,
-        Some("openclaw.plugin.json".to_owned()),
-        PluginCompatibilityMode::OpenClawModern,
-        Some(path),
-        primary_entry_path.as_deref(),
-        manifest,
-    );
-
-    Ok(descriptor)
-}
-
-fn parse_openclaw_legacy_package_descriptors(
-    path: &Path,
-    known_files: &BTreeSet<PathBuf>,
-) -> Result<Vec<PluginDescriptor>, IntegrationError> {
-    let document = parse_json_document::<OpenClawPackageJsonDocument>(path)?;
-    let Some(openclaw) = document.openclaw.as_ref() else {
-        return Ok(Vec::new());
-    };
-
-    let package_root = path.parent().unwrap_or(path);
-    let sibling_openclaw_manifest = package_root.join(OPENCLAW_PACKAGE_MANIFEST_FILE_NAME);
-    if known_files.contains(&sibling_openclaw_manifest) || sibling_openclaw_manifest.is_file() {
-        return Ok(Vec::new());
-    }
-
-    let extension_entries = resolve_openclaw_legacy_extension_entries(package_root, &document);
-    if extension_entries.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    let multiple_entries = extension_entries.len() > 1;
-    let setup_entry_path = openclaw
-        .setup_entry
-        .as_deref()
-        .and_then(|entry| resolve_openclaw_relative_path(package_root, entry));
-    let mut descriptors = Vec::new();
-
-    for entry_path in extension_entries {
-        let plugin_id = derive_openclaw_legacy_plugin_id(
-            document.name.as_deref(),
-            &entry_path,
-            multiple_entries,
-        );
-        let manifest = build_openclaw_legacy_manifest(
-            &document,
-            plugin_id,
-            &entry_path,
-            setup_entry_path.as_deref(),
-        );
-        descriptors.push(build_plugin_descriptor(
-            &entry_path,
-            PluginSourceKind::PackageManifest,
-            PluginContractDialect::OpenClawLegacyPackage,
-            Some("package.json#openclaw".to_owned()),
-            PluginCompatibilityMode::OpenClawLegacy,
-            Some(path),
-            Some(&entry_path),
-            manifest,
-        ));
-    }
-
-    Ok(descriptors)
-}
-
-fn parse_json_document<T>(path: &Path) -> Result<T, IntegrationError>
-where
-    T: for<'de> Deserialize<'de>,
-{
-    let content = read_utf8_file(path)?;
-    serde_json::from_str(content.trim()).map_err(|error| IntegrationError::PluginManifestParse {
-        path: path.display().to_string(),
-        reason: error.to_string(),
-    })
-}
-
-fn read_utf8_file(path: &Path) -> Result<String, IntegrationError> {
-    let bytes = fs::read(path).map_err(|error| IntegrationError::PluginFileRead {
-        path: path.display().to_string(),
-        reason: error.to_string(),
-    })?;
-
-    String::from_utf8(bytes).map_err(|error| IntegrationError::PluginManifestParse {
-        path: path.display().to_string(),
-        reason: error.to_string(),
-    })
-}
-
-fn validate_openclaw_manifest_document(
-    document: &OpenClawManifestDocument,
-    path: &Path,
-) -> Result<(), IntegrationError> {
-    if document.id.trim().is_empty() {
-        return Err(IntegrationError::PluginManifestParse {
-            path: path.display().to_string(),
-            reason: "openclaw.plugin.json must declare id".to_owned(),
-        });
-    }
-
-    if !matches!(document.config_schema.as_ref(), Some(Value::Object(_))) {
-        return Err(IntegrationError::PluginManifestParse {
-            path: path.display().to_string(),
-            reason: "openclaw.plugin.json must declare configSchema object".to_owned(),
-        });
-    }
-
-    Ok(())
-}
-
-fn build_openclaw_manifest(
-    document: &OpenClawManifestDocument,
-    package_document: Option<&OpenClawPackageJsonDocument>,
-    primary_entry_path: Option<&Path>,
-    setup_entry_path: Option<&Path>,
-    compatibility_mode: PluginCompatibilityMode,
-) -> PluginManifest {
-    let mut metadata = BTreeMap::new();
-
-    metadata.insert("bridge_kind".to_owned(), "process_stdio".to_owned());
-    metadata.insert(
-        "adapter_family".to_owned(),
-        match compatibility_mode {
-            PluginCompatibilityMode::Native => "native".to_owned(),
-            PluginCompatibilityMode::OpenClawModern => {
-                OPENCLAW_MODERN_COMPATIBILITY_ADAPTER_FAMILY.to_owned()
-            }
-            PluginCompatibilityMode::OpenClawLegacy => {
-                OPENCLAW_LEGACY_COMPATIBILITY_ADAPTER_FAMILY.to_owned()
-            }
-        },
-    );
-
-    if let Some(entry) = primary_entry_path {
-        metadata.insert("entrypoint".to_owned(), path_to_string(entry));
-    }
-    if let Some(setup_entry) = setup_entry_path {
-        metadata.insert("setup_entrypoint".to_owned(), path_to_string(setup_entry));
-    }
-    if let Some(kind) = normalize_optional_manifest_string(document.kind.clone()) {
-        metadata.insert("openclaw_kind".to_owned(), kind);
-    }
-    if let Some(package_document) = package_document {
-        if let Some(name) = normalize_optional_manifest_string(package_document.name.clone()) {
-            metadata.insert("openclaw_package_name".to_owned(), name);
-        }
-        if let Some(version) = normalize_optional_manifest_string(package_document.version.clone())
-        {
-            metadata.insert("openclaw_package_version".to_owned(), version);
-        }
-        if let Some(description) =
-            normalize_optional_manifest_string(package_document.description.clone())
-        {
-            metadata.insert("openclaw_package_description".to_owned(), description);
-        }
-        if let Some(channel) = package_document
-            .openclaw
-            .as_ref()
-            .and_then(|openclaw| openclaw.channel.as_ref())
-        {
-            if let Some(channel_id) = normalize_optional_manifest_string(channel.id.clone()) {
-                metadata.insert("openclaw_channel_id".to_owned(), channel_id);
-            }
-            if let Some(label) = normalize_optional_manifest_string(channel.label.clone()) {
-                metadata.insert("openclaw_channel_label".to_owned(), label);
-            }
-            if let Some(blurb) = normalize_optional_manifest_string(channel.blurb.clone()) {
-                metadata.insert("openclaw_channel_blurb".to_owned(), blurb);
-            }
-            if let Some(docs_path) = normalize_optional_manifest_string(channel.docs_path.clone()) {
-                metadata.insert("openclaw_channel_docs_path".to_owned(), docs_path);
-            }
-            let aliases = normalize_manifest_string_list(channel.aliases.clone());
-            if !aliases.is_empty()
-                && let Ok(encoded) = serde_json::to_string(&aliases)
-            {
-                metadata.insert("openclaw_channel_aliases_json".to_owned(), encoded);
-            }
-        }
-        if let Some(install) = package_document
-            .openclaw
-            .as_ref()
-            .and_then(|openclaw| openclaw.install.as_ref())
-        {
-            if let Some(npm_spec) = normalize_optional_manifest_string(install.npm_spec.clone()) {
-                metadata.insert("openclaw_install_npm_spec".to_owned(), npm_spec);
-            }
-            if let Some(local_path) = normalize_optional_manifest_string(install.local_path.clone())
-            {
-                metadata.insert("openclaw_install_local_path".to_owned(), local_path);
-            }
-            if let Some(min_host_version) =
-                normalize_optional_manifest_string(install.min_host_version.clone())
-            {
-                metadata.insert(
-                    "openclaw_install_min_host_version".to_owned(),
-                    min_host_version,
-                );
-            }
-        }
-    }
-
-    if !document.channels.is_empty()
-        && let Ok(encoded) =
-            serde_json::to_string(&normalize_manifest_string_list(document.channels.clone()))
-    {
-        metadata.insert("openclaw_channels_json".to_owned(), encoded);
-    }
-    if !document.providers.is_empty()
-        && let Ok(encoded) =
-            serde_json::to_string(&normalize_manifest_string_list(document.providers.clone()))
-    {
-        metadata.insert("openclaw_providers_json".to_owned(), encoded);
-    }
-    if !document.skills.is_empty()
-        && let Ok(encoded) =
-            serde_json::to_string(&normalize_manifest_string_list(document.skills.clone()))
-    {
-        metadata.insert("openclaw_skills_json".to_owned(), encoded);
-    }
-    if !document.provider_auth_env_vars.is_empty()
-        && let Ok(encoded) = serde_json::to_string(&document.provider_auth_env_vars)
-    {
-        metadata.insert("openclaw_provider_auth_env_vars_json".to_owned(), encoded);
-    }
-    if !document.provider_auth_choices.is_empty()
-        && let Ok(encoded) = serde_json::to_string(&document.provider_auth_choices)
-    {
-        metadata.insert("openclaw_provider_auth_choices_json".to_owned(), encoded);
-    }
-    if !document.ui_hints.is_empty()
-        && let Ok(encoded) = serde_json::to_string(&document.ui_hints)
-    {
-        metadata.insert("openclaw_ui_hints_json".to_owned(), encoded);
-    }
-    if document.enabled_by_default {
-        metadata.insert("openclaw_enabled_by_default".to_owned(), "true".to_owned());
-    }
-    if let Some(language) = primary_entry_path
-        .map(detect_language)
-        .filter(|language| language != "unknown")
-    {
-        metadata.insert(
-            "source_language".to_owned(),
-            normalize_language_name(&language),
-        );
-    }
-
-    normalize_plugin_manifest(PluginManifest {
-        api_version: Some(CURRENT_PLUGIN_MANIFEST_API_VERSION.to_owned()),
-        version: normalize_optional_manifest_string(document.version.clone()).or_else(|| {
-            package_document
-                .and_then(|package| normalize_optional_manifest_string(package.version.clone()))
-        }),
-        plugin_id: document.id.trim().to_owned(),
-        provider_id: document.id.trim().to_owned(),
-        connector_name: document.id.trim().to_owned(),
-        channel_id: None,
-        endpoint: None,
-        capabilities: derive_openclaw_capabilities(
-            document.providers.as_slice(),
-            document.channels.as_slice(),
-            document.skills.as_slice(),
-            document.kind.as_deref(),
-        ),
-        trust_tier: PluginTrustTier::default(),
-        metadata,
-        summary: normalize_optional_manifest_string(
-            document
-                .description
-                .clone()
-                .or_else(|| document.name.clone()),
-        ),
-        tags: derive_openclaw_tags(
-            compatibility_mode,
-            document.providers.as_slice(),
-            document.channels.as_slice(),
-            document.skills.as_slice(),
-            document.kind.as_deref(),
-        ),
-        input_examples: Vec::new(),
-        output_examples: Vec::new(),
-        defer_loading: setup_entry_path.is_some(),
-        setup: derive_openclaw_setup(document, setup_entry_path),
-        slot_claims: derive_openclaw_slot_claims(document.kind.as_deref()),
-        compatibility: None,
-    })
-}
-
-fn build_openclaw_legacy_manifest(
-    package_document: &OpenClawPackageJsonDocument,
-    plugin_id: String,
-    primary_entry_path: &Path,
-    setup_entry_path: Option<&Path>,
-) -> PluginManifest {
-    let synthetic_document = OpenClawManifestDocument {
-        id: plugin_id,
-        config_schema: Some(Value::Object(Default::default())),
-        enabled_by_default: false,
-        kind: None,
-        channels: Vec::new(),
-        providers: Vec::new(),
-        provider_auth_env_vars: BTreeMap::new(),
-        provider_auth_choices: Vec::new(),
-        skills: Vec::new(),
-        name: package_document.name.clone(),
-        description: package_document.description.clone(),
-        version: package_document.version.clone(),
-        ui_hints: BTreeMap::new(),
-    };
-
-    let mut manifest = build_openclaw_manifest(
-        &synthetic_document,
-        Some(package_document),
-        Some(primary_entry_path),
-        setup_entry_path,
-        PluginCompatibilityMode::OpenClawLegacy,
-    );
-    manifest
-        .metadata
-        .insert("openclaw_legacy_package".to_owned(), "true".to_owned());
-    manifest.summary = normalize_optional_manifest_string(
-        package_document
-            .description
-            .clone()
-            .or_else(|| package_document.name.clone()),
-    );
-    manifest
-}
-
-fn resolve_openclaw_primary_entry_path(
-    package_root: &Path,
-    package_document: Option<&OpenClawPackageJsonDocument>,
-    prefer_declared_extension: bool,
-) -> Option<PathBuf> {
-    if prefer_declared_extension && let Some(package_document) = package_document {
-        let entries = resolve_openclaw_extension_entries(package_root, package_document);
-        if let Some(first) = entries.into_iter().next() {
-            return Some(first);
-        }
-    }
-
-    resolve_openclaw_default_entry_path(package_root)
-}
-
-fn resolve_openclaw_legacy_extension_entries(
-    package_root: &Path,
-    package_document: &OpenClawPackageJsonDocument,
-) -> Vec<PathBuf> {
-    let declared = resolve_openclaw_extension_entries(package_root, package_document);
-    if !declared.is_empty() {
-        return declared;
-    }
-
-    resolve_openclaw_default_entry_path(package_root)
-        .into_iter()
-        .collect()
-}
-
-fn resolve_openclaw_extension_entries(
-    package_root: &Path,
-    package_document: &OpenClawPackageJsonDocument,
-) -> Vec<PathBuf> {
-    package_document
-        .openclaw
-        .as_ref()
-        .map(|metadata| metadata.extensions.as_slice())
-        .unwrap_or_default()
-        .iter()
-        .filter_map(|entry| resolve_openclaw_relative_path(package_root, entry))
-        .collect()
-}
-
-fn resolve_openclaw_default_entry_path(package_root: &Path) -> Option<PathBuf> {
-    for candidate in ["index.ts", "index.js", "index.mjs", "index.cjs"] {
-        let entry = package_root.join(candidate);
-        if entry.is_file() {
-            return Some(entry);
-        }
-    }
-
-    None
-}
-
-fn resolve_openclaw_relative_path(package_root: &Path, raw: &str) -> Option<PathBuf> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-
-    let candidate = package_root.join(trimmed);
-    Some(candidate)
-}
-
-fn derive_openclaw_legacy_plugin_id(
-    package_name: Option<&str>,
-    entry_path: &Path,
-    has_multiple_extensions: bool,
-) -> String {
-    let base = entry_path
-        .file_stem()
-        .and_then(|stem| stem.to_str())
-        .map(str::trim)
-        .filter(|stem| !stem.is_empty())
-        .unwrap_or("plugin");
-
-    let Some(package_name) = package_name.map(str::trim).filter(|name| !name.is_empty()) else {
-        return base.to_owned();
-    };
-
-    let unscoped = package_name.rsplit('/').next().unwrap_or(package_name);
-    let canonical = unscoped
-        .strip_suffix("-provider")
-        .unwrap_or(unscoped)
-        .trim();
-
-    if !has_multiple_extensions {
-        return canonical.to_owned();
-    }
-
-    format!("{canonical}/{base}")
-}
-
-fn derive_openclaw_capabilities(
-    providers: &[String],
-    channels: &[String],
-    skills: &[String],
-    kind: Option<&str>,
-) -> BTreeSet<Capability> {
-    let mut capabilities = BTreeSet::new();
-    if !providers.is_empty() || !channels.is_empty() {
-        capabilities.insert(Capability::InvokeConnector);
-    }
-    if !skills.is_empty() {
-        capabilities.insert(Capability::InvokeTool);
-    }
-
-    match kind.map(|value| value.trim().to_ascii_lowercase()) {
-        Some(kind) if kind == "memory" => {
-            capabilities.insert(Capability::MemoryRead);
-            capabilities.insert(Capability::MemoryWrite);
-        }
-        Some(kind) if kind == "context-engine" => {
-            capabilities.insert(Capability::ObserveTelemetry);
-        }
-        _ => {}
-    }
-
-    capabilities
-}
-
-fn derive_openclaw_tags(
-    compatibility_mode: PluginCompatibilityMode,
-    providers: &[String],
-    channels: &[String],
-    skills: &[String],
-    kind: Option<&str>,
-) -> Vec<String> {
-    let mut tags = vec![
-        "openclaw".to_owned(),
-        compatibility_mode.as_str().to_owned(),
-        "compat".to_owned(),
-    ];
-    if !providers.is_empty() {
-        tags.push("provider".to_owned());
-    }
-    if !channels.is_empty() {
-        tags.push("channel".to_owned());
-    }
-    if !skills.is_empty() {
-        tags.push("skill".to_owned());
-    }
-    if let Some(kind) = kind.map(str::trim).filter(|kind| !kind.is_empty()) {
-        tags.push(kind.to_ascii_lowercase());
-    }
-
-    normalize_manifest_string_list(tags)
-}
-
-fn derive_openclaw_setup(
-    document: &OpenClawManifestDocument,
-    setup_entry_path: Option<&Path>,
-) -> Option<PluginSetup> {
-    let required_env_vars = document
-        .provider_auth_env_vars
-        .values()
-        .flat_map(|values| values.iter().cloned())
-        .collect::<Vec<_>>();
-    let docs_urls = document
-        .provider_auth_choices
-        .iter()
-        .filter_map(|choice| choice.get("docsUrl"))
-        .filter_map(Value::as_str)
-        .map(str::to_owned)
-        .collect::<Vec<_>>();
-    let surface = if !document.channels.is_empty() {
-        Some("channel".to_owned())
-    } else if !document.providers.is_empty() {
-        Some("provider".to_owned())
-    } else if !document.skills.is_empty() {
-        Some("skill".to_owned())
-    } else {
-        Some("plugin".to_owned())
-    };
-    let remediation = Some(
-        "enable the required OpenClaw compatibility shim and configure plugin settings before activation"
-            .to_owned(),
-    );
-    let setup = PluginSetup {
-        mode: if setup_entry_path.is_some() {
-            PluginSetupMode::GovernedEntry
-        } else {
-            PluginSetupMode::MetadataOnly
-        },
-        surface,
-        required_env_vars: normalize_manifest_string_list(required_env_vars),
-        recommended_env_vars: Vec::new(),
-        required_config_keys: vec![format!("plugins.entries.{}", document.id.trim())],
-        default_env_var: document
-            .provider_auth_env_vars
-            .values()
-            .flat_map(|values| values.iter())
-            .next()
-            .cloned(),
-        docs_urls: normalize_manifest_string_list(docs_urls),
-        remediation,
-    };
-
-    (!setup.is_effectively_empty()).then_some(setup.normalized())
-}
-
-fn derive_openclaw_slot_claims(kind: Option<&str>) -> Vec<PluginSlotClaim> {
-    match kind.map(|value| value.trim().to_ascii_lowercase()) {
-        Some(kind) if kind == "memory" => vec![PluginSlotClaim {
-            slot: "openclaw_kind".to_owned(),
-            key: "memory".to_owned(),
-            mode: PluginSlotMode::Exclusive,
-        }],
-        Some(kind) if kind == "context-engine" => vec![PluginSlotClaim {
-            slot: "openclaw_kind".to_owned(),
-            key: "context_engine".to_owned(),
-            mode: PluginSlotMode::Exclusive,
-        }],
-        _ => Vec::new(),
-    }
 }
 
 fn validate_package_manifest_document_contract(
@@ -2686,1568 +2035,4 @@ struct ManifestFieldConflict {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    fn unique_tmp_dir(prefix: &str) -> PathBuf {
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("clock should be monotonic")
-            .as_nanos();
-        std::env::temp_dir().join(format!("{}-{}", prefix, nanos))
-    }
-
-    fn sample_pack() -> VerticalPackManifest {
-        VerticalPackManifest {
-            pack_id: "sample-pack".to_owned(),
-            domain: "engineering".to_owned(),
-            version: "0.1.0".to_owned(),
-            default_route: crate::contracts::ExecutionRoute {
-                harness_kind: crate::contracts::HarnessKind::EmbeddedPi,
-                adapter: Some("pi-local".to_owned()),
-            },
-            allowed_connectors: BTreeSet::new(),
-            granted_capabilities: BTreeSet::new(),
-            metadata: BTreeMap::new(),
-        }
-    }
-
-    fn scan_diagnostic<'a>(
-        report: &'a PluginScanReport,
-        code: PluginDiagnosticCode,
-        plugin_id: &str,
-    ) -> Option<&'a PluginDiagnosticFinding> {
-        report
-            .diagnostic_findings
-            .iter()
-            .find(|finding| finding.code == code && finding.plugin_id.as_deref() == Some(plugin_id))
-    }
-
-    #[test]
-    fn scanner_finds_manifest_in_rust_and_python_files() {
-        let root = unique_tmp_dir("loong-plugin-scan");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let rust_file = root.join("openrouter.rs");
-        fs::write(
-            &rust_file,
-            r#"
-// LOONG_PLUGIN_START
-// {
-//   "plugin_id": "openrouter-rs",
-//   "provider_id": "openrouter",
-//   "connector_name": "openrouter",
-//   "channel_id": "primary",
-//   "endpoint": "https://openrouter.ai/api/v1/chat/completions",
-//   "capabilities": ["InvokeConnector", "ObserveTelemetry"],
-//   "metadata": {"version":"0.2.0","lang":"rust"}
-// }
-// LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write rust plugin");
-
-        let py_file = root.join("slack_plugin.py");
-        fs::write(
-            &py_file,
-            r#"
-# LOONG_PLUGIN_START
-# {
-#   "plugin_id": "slack-py",
-#   "provider_id": "slack",
-#   "connector_name": "slack",
-#   "channel_id": "alerts",
-#   "endpoint": "https://hooks.slack.com/services/aaa/bbb/ccc",
-#   "capabilities": ["InvokeConnector"],
-#   "metadata": {"version":"1.1.0","lang":"python"}
-# }
-# LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write python plugin");
-
-        let scanner = PluginScanner::new();
-        let report = scanner.scan_path(&root).expect("scan should succeed");
-        assert_eq!(report.matched_plugins, 2);
-        assert!(
-            report
-                .descriptors
-                .iter()
-                .any(|descriptor| descriptor.manifest.provider_id == "openrouter")
-        );
-        assert!(
-            report
-                .descriptors
-                .iter()
-                .all(|descriptor| descriptor.source_kind == PluginSourceKind::EmbeddedSource)
-        );
-        assert!(
-            report
-                .descriptors
-                .iter()
-                .all(|descriptor| descriptor.package_manifest_path.is_none())
-        );
-        assert!(report.descriptors.iter().all(|descriptor| matches!(
-            descriptor.manifest.trust_tier,
-            PluginTrustTier::Unverified
-        )));
-        assert!(
-            report
-                .descriptors
-                .iter()
-                .any(|descriptor| descriptor.manifest.provider_id == "slack")
-        );
-        assert_eq!(
-            report
-                .diagnostic_findings
-                .iter()
-                .filter(|finding| finding.code == PluginDiagnosticCode::EmbeddedSourceLegacyContract)
-                .count(),
-            2
-        );
-        assert_eq!(
-            report
-                .diagnostic_findings
-                .iter()
-                .filter(|finding| finding.code == PluginDiagnosticCode::LegacyMetadataVersion)
-                .count(),
-            2
-        );
-    }
-
-    #[test]
-    fn scanner_finds_package_manifest_file() {
-        let root = unique_tmp_dir("loong-plugin-package-manifest");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let manifest_file = root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "plugin_id": "tavily-search",
-  "version": "0.3.0",
-  "provider_id": "tavily",
-  "connector_name": "tavily-http",
-  "endpoint": "https://api.tavily.com/search",
-  "capabilities": ["InvokeConnector"],
-  "trust_tier": "verified-community",
-  "metadata": {
-    "bridge_kind": "http_json",
-    "adapter_family": "web-search"
-  },
-  "summary": "Manifest-discovered Tavily package",
-  "tags": ["search", "provider"],
-  "setup": {
-    "mode": "metadata_only",
-    "surface": " web_search ",
-    "required_env_vars": ["TAVILY_API_KEY", " ", "TAVILY_API_KEY"],
-    "recommended_env_vars": ["TEAM_TAVILY_KEY"],
-    "required_config_keys": ["tools.web_search.default_provider"],
-    "default_env_var": " TAVILY_API_KEY ",
-    "docs_urls": ["https://docs.example.com/tavily", "https://docs.example.com/tavily"],
-    "remediation": " set a Tavily credential before enabling search "
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let scanner = PluginScanner::new();
-        let report = scanner.scan_path(&root).expect("scan should succeed");
-
-        assert_eq!(report.scanned_files, 1);
-        assert_eq!(report.matched_plugins, 1);
-        assert_eq!(report.descriptors.len(), 1);
-        assert_eq!(
-            report.descriptors[0].path,
-            manifest_file.display().to_string()
-        );
-        assert_eq!(report.descriptors[0].language, "manifest");
-        assert_eq!(
-            report.descriptors[0].manifest.api_version.as_deref(),
-            Some(CURRENT_PLUGIN_MANIFEST_API_VERSION)
-        );
-        assert_eq!(
-            report.descriptors[0].manifest.version.as_deref(),
-            Some("0.3.0")
-        );
-        assert_eq!(report.descriptors[0].manifest.plugin_id, "tavily-search");
-        assert_eq!(report.descriptors[0].manifest.provider_id, "tavily");
-        assert_eq!(
-            report.descriptors[0]
-                .manifest
-                .metadata
-                .get("version")
-                .map(String::as_str),
-            Some("0.3.0")
-        );
-        assert_eq!(
-            report.descriptors[0].source_kind,
-            PluginSourceKind::PackageManifest
-        );
-        assert_eq!(
-            report.descriptors[0].package_root,
-            root.display().to_string()
-        );
-        assert_eq!(
-            report.descriptors[0].package_manifest_path,
-            Some(manifest_file.display().to_string())
-        );
-        assert_eq!(
-            report.descriptors[0].manifest.trust_tier,
-            PluginTrustTier::VerifiedCommunity
-        );
-        assert_eq!(
-            report.descriptors[0].manifest.setup,
-            Some(PluginSetup {
-                mode: PluginSetupMode::MetadataOnly,
-                surface: Some("web_search".to_owned()),
-                required_env_vars: vec!["TAVILY_API_KEY".to_owned()],
-                recommended_env_vars: vec!["TEAM_TAVILY_KEY".to_owned()],
-                required_config_keys: vec!["tools.web_search.default_provider".to_owned()],
-                default_env_var: Some("TAVILY_API_KEY".to_owned()),
-                docs_urls: vec!["https://docs.example.com/tavily".to_owned()],
-                remediation: Some("set a Tavily credential before enabling search".to_owned()),
-            })
-        );
-    }
-
-    #[test]
-    fn scanner_requires_api_version_for_package_manifest() {
-        let root = unique_tmp_dir("loong-plugin-package-api-required");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let manifest_file = root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "version": "1.0.0",
-  "plugin_id": "missing-api-version",
-  "provider_id": "missing-api-version",
-  "connector_name": "missing-api-version",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let error = PluginScanner::new()
-            .scan_path(&root)
-            .expect_err("package manifests must declare api_version");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("api_version"));
-        assert!(rendered.contains("package manifest"));
-    }
-
-    #[test]
-    fn scanner_requires_top_level_version_for_package_manifest() {
-        let root = unique_tmp_dir("loong-plugin-package-version-required");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let manifest_file = root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "plugin_id": "missing-version",
-  "provider_id": "missing-version",
-  "connector_name": "missing-version",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let error = PluginScanner::new()
-            .scan_path(&root)
-            .expect_err("package manifests must declare top-level version");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("top-level version"));
-        assert!(rendered.contains("package manifest"));
-    }
-
-    #[test]
-    fn scanner_rejects_legacy_version_metadata_in_package_manifest() {
-        let root = unique_tmp_dir("loong-plugin-package-legacy-version");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let manifest_file = root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.2.3",
-  "plugin_id": "legacy-version-metadata",
-  "provider_id": "legacy-version-metadata",
-  "connector_name": "legacy-version-metadata",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json",
-    "version": "1.2.3"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let error = PluginScanner::new()
-            .scan_path(&root)
-            .expect_err("package manifests should reject metadata.version");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("metadata.version"));
-        assert!(rendered.contains("top-level `version`"));
-    }
-
-    #[test]
-    fn scanner_rejects_reserved_metadata_namespace_in_package_manifest() {
-        let root = unique_tmp_dir("loong-plugin-package-reserved-metadata");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let manifest_file = root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.2.3",
-  "plugin_id": "reserved-metadata",
-  "provider_id": "reserved-metadata",
-  "connector_name": "reserved-metadata",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json",
-    "plugin_version": "1.2.3"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let error = PluginScanner::new()
-            .scan_path(&root)
-            .expect_err("package manifests should reject reserved metadata namespace");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("plugin_version"));
-        assert!(rendered.contains("reserved"));
-    }
-
-    #[test]
-    fn scanner_rejects_invalid_top_level_plugin_version() {
-        let root = unique_tmp_dir("loong-plugin-invalid-version");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let manifest_file = root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "not-a-semver",
-  "plugin_id": "bad-version",
-  "provider_id": "bad-version",
-  "connector_name": "bad-version",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let error = PluginScanner::new()
-            .scan_path(&root)
-            .expect_err("invalid plugin version should fail parse");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("invalid semver"));
-        assert!(rendered.contains("not-a-semver"));
-    }
-
-    #[test]
-    fn scanner_rejects_conflicting_top_level_and_metadata_version_in_source_manifest() {
-        let root = unique_tmp_dir("loong-plugin-source-version-conflict");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let source_file = root.join("plugin.py");
-        fs::write(
-            &source_file,
-            r#"
-# LOONG_PLUGIN_START
-# {
-#   "version": "1.2.3",
-#   "plugin_id": "source-version-conflict",
-#   "provider_id": "source-version-conflict",
-#   "connector_name": "source-version-conflict",
-#   "capabilities": ["InvokeConnector"],
-#   "metadata": {
-#     "bridge_kind": "http_json",
-#     "version": "9.9.9"
-#   }
-# }
-# LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write source manifest");
-
-        let error = PluginScanner::new()
-            .scan_path(&root)
-            .expect_err("source manifests should reject conflicting version truth");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("plugin version conflict"));
-        assert!(rendered.contains("1.2.3"));
-        assert!(rendered.contains("9.9.9"));
-    }
-
-    #[test]
-    fn scanner_rejects_unknown_package_manifest_fields() {
-        let root = unique_tmp_dir("loong-plugin-unknown-package-field");
-        fs::create_dir_all(&root).expect("create temp root");
-
-        let manifest_file = root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.0.0",
-  "plugin_id": "unknown-field",
-  "provider_id": "unknown-field",
-  "connector_name": "unknown-field",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  },
-  "slot_claim": []
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let error = PluginScanner::new()
-            .scan_path(&root)
-            .expect_err("unknown package manifest fields should fail parse");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("unknown field"));
-        assert!(rendered.contains("slot_claim"));
-    }
-
-    #[test]
-    fn scanner_prefers_package_manifest_over_embedded_source_manifest() {
-        let root = unique_tmp_dir("loong-plugin-precedence");
-        let package_root = root.join("pkg");
-        fs::create_dir_all(&package_root).expect("create temp root");
-
-        let manifest_file = package_root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.0.0",
-  "plugin_id": "package-plugin",
-  "provider_id": "package-provider",
-  "connector_name": "package-connector",
-  "channel_id": "package-channel",
-  "endpoint": "https://package.example/invoke",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let source_file = package_root.join("plugin.py");
-        fs::write(
-            &source_file,
-            r#"
-# LOONG_PLUGIN_START
-# {
-#   "plugin_id": "package-plugin",
-#   "provider_id": "package-provider",
-#   "connector_name": "package-connector",
-#   "channel_id": "package-channel",
-#   "endpoint": "https://package.example/invoke",
-#   "capabilities": ["InvokeConnector"],
-#   "metadata": {"bridge_kind":"http_json"}
-# }
-# LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write source plugin");
-
-        let scanner = PluginScanner::new();
-        let report = scanner.scan_path(&root).expect("scan should succeed");
-
-        assert_eq!(report.scanned_files, 2);
-        assert_eq!(report.matched_plugins, 1);
-        assert_eq!(report.descriptors.len(), 1);
-        assert_eq!(
-            report.descriptors[0].path,
-            manifest_file.display().to_string()
-        );
-        assert_eq!(
-            report.descriptors[0].source_kind,
-            PluginSourceKind::PackageManifest
-        );
-        assert_eq!(
-            report.descriptors[0].package_root,
-            package_root.display().to_string()
-        );
-        assert_eq!(
-            report.descriptors[0].package_manifest_path,
-            Some(manifest_file.display().to_string())
-        );
-        assert_eq!(report.descriptors[0].manifest.plugin_id, "package-plugin");
-        assert_eq!(
-            report.descriptors[0].manifest.provider_id,
-            "package-provider"
-        );
-        let finding = scan_diagnostic(
-            &report,
-            PluginDiagnosticCode::ShadowedEmbeddedSource,
-            "package-plugin",
-        )
-        .expect("shadowed embedded source finding");
-        assert_eq!(finding.phase, PluginDiagnosticPhase::Scan);
-        assert!(!finding.blocking);
-    }
-
-    #[test]
-    fn scanner_fails_when_package_manifest_conflicts_with_source_manifest() {
-        let root = unique_tmp_dir("loong-plugin-conflict");
-        let package_root = root.join("pkg");
-        fs::create_dir_all(&package_root).expect("create temp root");
-
-        let manifest_file = package_root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.0.0",
-  "plugin_id": "package-plugin",
-  "provider_id": "package-provider",
-  "connector_name": "package-connector",
-  "channel_id": "package-channel",
-  "endpoint": "https://package.example/invoke",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let source_file = package_root.join("plugin.py");
-        fs::write(
-            &source_file,
-            r#"
-# LOONG_PLUGIN_START
-# {
-#   "plugin_id": "package-plugin",
-#   "provider_id": "source-provider",
-#   "connector_name": "package-connector",
-#   "channel_id": "package-channel",
-#   "endpoint": "https://package.example/invoke",
-#   "capabilities": ["InvokeConnector"],
-#   "metadata": {"bridge_kind":"http_json"}
-# }
-# LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write source plugin");
-
-        let scanner = PluginScanner::new();
-        let error = scanner
-            .scan_path(&root)
-            .expect_err("conflicting manifests should fail");
-
-        assert_eq!(
-            error,
-            IntegrationError::PluginManifestConflict {
-                package_manifest_path: manifest_file.display().to_string(),
-                source_path: source_file.display().to_string(),
-                field: "provider_id".to_owned(),
-                package_value: "\"package-provider\"".to_owned(),
-                source_value: "\"source-provider\"".to_owned(),
-            }
-        );
-    }
-
-    #[test]
-    fn scanner_uses_nearest_package_manifest_for_nested_package_roots() {
-        let root = unique_tmp_dir("loong-plugin-nested-package-root");
-        let outer_root = root.join("outer");
-        let inner_root = outer_root.join("inner");
-        fs::create_dir_all(&inner_root).expect("create nested root");
-
-        let outer_manifest_file = outer_root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &outer_manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.0.0",
-  "plugin_id": "outer-plugin",
-  "provider_id": "outer-provider",
-  "connector_name": "outer-connector",
-  "channel_id": "outer-channel",
-  "endpoint": "https://outer.example/invoke",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write outer package manifest");
-
-        let inner_manifest_file = inner_root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &inner_manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.0.0",
-  "plugin_id": "inner-plugin",
-  "provider_id": "inner-provider",
-  "connector_name": "inner-connector",
-  "channel_id": "inner-channel",
-  "endpoint": "https://inner.example/invoke",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write inner package manifest");
-
-        let source_file = inner_root.join("plugin.py");
-        fs::write(
-            &source_file,
-            r#"
-# LOONG_PLUGIN_START
-# {
-#   "plugin_id": "inner-plugin",
-#   "provider_id": "inner-provider",
-#   "connector_name": "inner-connector",
-#   "channel_id": "inner-channel",
-#   "endpoint": "https://inner.example/invoke",
-#   "capabilities": ["InvokeConnector"],
-#   "metadata": {"bridge_kind":"http_json"}
-# }
-# LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write nested source plugin");
-
-        let scanner = PluginScanner::new();
-        let report = scanner.scan_path(&root).expect("scan should succeed");
-
-        assert_eq!(report.matched_plugins, 2);
-        assert_eq!(report.descriptors.len(), 2);
-        assert!(
-            report
-                .descriptors
-                .iter()
-                .any(|descriptor| descriptor.path == outer_manifest_file.display().to_string())
-        );
-        assert!(
-            report
-                .descriptors
-                .iter()
-                .any(|descriptor| descriptor.path == inner_manifest_file.display().to_string())
-        );
-    }
-
-    #[test]
-    fn scanner_allows_source_only_optional_fields_under_package_manifest() {
-        let root = unique_tmp_dir("loong-plugin-optional-source-fields");
-        let package_root = root.join("pkg");
-        fs::create_dir_all(&package_root).expect("create temp root");
-
-        let manifest_file = package_root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.0.0",
-  "plugin_id": "package-plugin",
-  "provider_id": "package-provider",
-  "connector_name": "package-connector",
-  "channel_id": "package-channel",
-  "endpoint": "https://package.example/invoke",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let source_file = package_root.join("plugin.py");
-        fs::write(
-            &source_file,
-            r#"
-# LOONG_PLUGIN_START
-# {
-#   "plugin_id": "package-plugin",
-#   "provider_id": "package-provider",
-#   "connector_name": "package-connector",
-#   "channel_id": "package-channel",
-#   "endpoint": "https://package.example/invoke",
-#   "capabilities": ["InvokeConnector"],
-#   "metadata": {"bridge_kind":"http_json","legacy_source":"true"},
-#   "summary": "legacy source summary",
-#   "tags": ["legacy", "source"],
-#   "input_examples": [{"query":"hello"}]
-# }
-# LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write source plugin");
-
-        let scanner = PluginScanner::new();
-        let report = scanner.scan_path(&root).expect("scan should succeed");
-
-        assert_eq!(report.scanned_files, 2);
-        assert_eq!(report.matched_plugins, 1);
-        assert_eq!(report.descriptors.len(), 1);
-        assert_eq!(
-            report.descriptors[0].path,
-            manifest_file.display().to_string()
-        );
-        assert_eq!(report.descriptors[0].manifest.summary, None);
-        assert!(report.descriptors[0].manifest.tags.is_empty());
-        assert!(report.descriptors[0].manifest.input_examples.is_empty());
-        assert!(
-            !report.descriptors[0]
-                .manifest
-                .metadata
-                .contains_key("legacy_source")
-        );
-        assert_eq!(
-            report.descriptors[0].manifest.provider_id,
-            "package-provider"
-        );
-        assert_eq!(report.descriptors[0].language, "manifest");
-        let finding = scan_diagnostic(
-            &report,
-            PluginDiagnosticCode::ShadowedEmbeddedSource,
-            "package-plugin",
-        )
-        .expect("shadowed embedded source finding");
-        assert_eq!(finding.phase, PluginDiagnosticPhase::Scan);
-        assert!(!finding.blocking);
-    }
-
-    #[test]
-    fn scanner_falls_back_to_embedded_source_manifest_without_package_manifest() {
-        let root = unique_tmp_dir("loong-plugin-source-fallback");
-        let package_root = root.join("pkg");
-        fs::create_dir_all(&package_root).expect("create temp root");
-
-        let source_file = package_root.join("plugin.py");
-        fs::write(
-            &source_file,
-            r#"
-# LOONG_PLUGIN_START
-# {
-#   "plugin_id": "source-plugin",
-#   "provider_id": "source-provider",
-#   "connector_name": "source-connector",
-#   "channel_id": "source-channel",
-#   "endpoint": "https://source.example/invoke",
-#   "capabilities": ["InvokeConnector"],
-#   "metadata": {"bridge_kind":"process_stdio"},
-#   "setup": {
-#     "surface": "channel",
-#     "required_env_vars": ["SOURCE_TOKEN"],
-#     "default_env_var": "SOURCE_TOKEN"
-#   }
-# }
-# LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write source plugin");
-
-        let scanner = PluginScanner::new();
-        let report = scanner.scan_path(&root).expect("scan should succeed");
-
-        assert_eq!(report.scanned_files, 1);
-        assert_eq!(report.matched_plugins, 1);
-        assert_eq!(report.descriptors.len(), 1);
-        assert_eq!(
-            report.descriptors[0].path,
-            source_file.display().to_string()
-        );
-        assert_eq!(
-            report.descriptors[0].source_kind,
-            PluginSourceKind::EmbeddedSource
-        );
-        assert_eq!(
-            report.descriptors[0].package_root,
-            package_root.display().to_string()
-        );
-        assert_eq!(report.descriptors[0].package_manifest_path, None);
-        assert_eq!(report.descriptors[0].language, "py");
-        assert_eq!(report.descriptors[0].manifest.plugin_id, "source-plugin");
-        assert_eq!(
-            report.descriptors[0].manifest.provider_id,
-            "source-provider"
-        );
-        let finding = scan_diagnostic(
-            &report,
-            PluginDiagnosticCode::EmbeddedSourceLegacyContract,
-            "source-plugin",
-        )
-        .expect("embedded source legacy finding");
-        assert_eq!(finding.phase, PluginDiagnosticPhase::Scan);
-        assert!(!finding.blocking);
-        assert_eq!(
-            report.descriptors[0].manifest.setup,
-            Some(PluginSetup {
-                mode: PluginSetupMode::MetadataOnly,
-                surface: Some("channel".to_owned()),
-                required_env_vars: vec!["SOURCE_TOKEN".to_owned()],
-                recommended_env_vars: Vec::new(),
-                required_config_keys: Vec::new(),
-                default_env_var: Some("SOURCE_TOKEN".to_owned()),
-                docs_urls: Vec::new(),
-                remediation: None,
-            })
-        );
-    }
-
-    #[test]
-    fn scanner_treats_empty_metadata_only_setup_as_absent() {
-        let root = unique_tmp_dir("loong-plugin-empty-setup");
-        let package_root = root.join("pkg");
-        fs::create_dir_all(&package_root).expect("create temp root");
-
-        let manifest_file = package_root.join(PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &manifest_file,
-            r#"
-{
-  "api_version": "v1alpha1",
-  "version": "1.0.0",
-  "plugin_id": "package-plugin",
-  "provider_id": "package-provider",
-  "connector_name": "package-connector",
-  "channel_id": "package-channel",
-  "endpoint": "https://package.example/invoke",
-  "capabilities": ["InvokeConnector"],
-  "metadata": {
-    "bridge_kind": "http_json"
-  }
-}
-"#,
-        )
-        .expect("write package manifest");
-
-        let source_file = package_root.join("plugin.py");
-        fs::write(
-            &source_file,
-            r#"
-# LOONG_PLUGIN_START
-# {
-#   "plugin_id": "package-plugin",
-#   "provider_id": "package-provider",
-#   "connector_name": "package-connector",
-#   "channel_id": "package-channel",
-#   "endpoint": "https://package.example/invoke",
-#   "capabilities": ["InvokeConnector"],
-#   "metadata": {"bridge_kind":"http_json"},
-#   "setup": {}
-# }
-# LOONG_PLUGIN_END
-"#,
-        )
-        .expect("write source plugin");
-
-        let scanner = PluginScanner::new();
-        let report = scanner.scan_path(&root).expect("scan should succeed");
-
-        assert_eq!(report.scanned_files, 2);
-        assert_eq!(report.matched_plugins, 1);
-        assert_eq!(report.descriptors.len(), 1);
-        assert_eq!(report.descriptors[0].manifest.setup, None);
-    }
-
-    #[test]
-    fn scanner_recognizes_openclaw_modern_manifest_through_explicit_compatibility_boundary() {
-        let root = unique_tmp_dir("loong-openclaw-modern");
-        let package_root = root.join("pkg");
-        fs::create_dir_all(package_root.join("dist")).expect("create temp root");
-
-        let package_manifest = package_root.join(OPENCLAW_PACKAGE_MANIFEST_FILE_NAME);
-        fs::write(
-            &package_manifest,
-            r#"
-{
-  "id": "search-sdk",
-  "name": "Search SDK",
-  "description": "OpenClaw search integration",
-  "version": "1.2.3",
-  "kind": "provider",
-  "providers": ["web_search"],
-  "channels": ["search"],
-  "skills": ["search"],
-  "configSchema": {}
-}
-"#,
-        )
-        .expect("write openclaw manifest");
-
-        let package_json = package_root.join(PACKAGE_JSON_FILE_NAME);
-        fs::write(
-            &package_json,
-            r#"
-{
-  "name": "@acme/search-provider",
-  "version": "1.2.3",
-  "description": "Search provider package",
-  "openclaw": {
-    "extensions": ["dist/index.js"],
-    "setupEntry": "dist/setup.js",
-    "channel": {
-      "id": "search",
-      "label": "Search",
-      "aliases": ["web-search"]
-    }
-  }
-}
-"#,
-        )
-        .expect("write package.json");
-        fs::write(package_root.join("dist/index.js"), "export {};\n").expect("write entry");
-        fs::write(package_root.join("dist/setup.js"), "export {};\n").expect("write setup");
-
-        let report = PluginScanner::new()
-            .scan_path(&root)
-            .expect("scan should succeed");
-
-        assert_eq!(report.matched_plugins, 1);
-        assert_eq!(report.descriptors.len(), 1);
-        assert_eq!(
-            report.descriptors[0].dialect,
-            PluginContractDialect::OpenClawModernManifest
-        );
-        assert_eq!(
-            report.descriptors[0].compatibility_mode,
-            PluginCompatibilityMode::OpenClawModern
-        );
-        assert_eq!(report.descriptors[0].language, "js");
-        assert_eq!(report.descriptors[0].manifest.plugin_id, "search-sdk");
-        assert_eq!(
-            report.descriptors[0].package_manifest_path,
-            Some(package_manifest.display().to_string())
-        );
-        assert_eq!(
-            report.descriptors[0].path,
-            package_root.join("dist/index.js").display().to_string()
-        );
-        let foreign = scan_diagnostic(
-            &report,
-            PluginDiagnosticCode::ForeignDialectContract,
-            "search-sdk",
-        )
-        .expect("foreign dialect diagnostic");
-        assert_eq!(foreign.phase, PluginDiagnosticPhase::Scan);
-        assert!(!foreign.blocking);
-        assert_eq!(
-            report.descriptors[0]
-                .manifest
-                .metadata
-                .get("adapter_family")
-                .map(String::as_str),
-            Some(OPENCLAW_MODERN_COMPATIBILITY_ADAPTER_FAMILY)
-        );
-    }
-
-    #[test]
-    fn scanner_recognizes_openclaw_legacy_package_metadata_without_promoting_it_to_native() {
-        let root = unique_tmp_dir("loong-openclaw-legacy");
-        let package_root = root.join("pkg");
-        fs::create_dir_all(package_root.join("dist")).expect("create temp root");
-
-        let package_json = package_root.join(PACKAGE_JSON_FILE_NAME);
-        fs::write(
-            &package_json,
-            r#"
-{
-  "name": "@acme/search-provider",
-  "version": "0.9.0",
-  "description": "Legacy OpenClaw package",
-  "openclaw": {
-    "extensions": ["dist/index.js"],
-    "setupEntry": "dist/setup.js"
-  }
-}
-"#,
-        )
-        .expect("write legacy package.json");
-        fs::write(package_root.join("dist/index.js"), "export {};\n").expect("write entry");
-        fs::write(package_root.join("dist/setup.js"), "export {};\n").expect("write setup");
-
-        let report = PluginScanner::new()
-            .scan_path(&root)
-            .expect("scan should succeed");
-
-        assert_eq!(report.matched_plugins, 1);
-        assert_eq!(report.descriptors.len(), 1);
-        assert_eq!(
-            report.descriptors[0].dialect,
-            PluginContractDialect::OpenClawLegacyPackage
-        );
-        assert_eq!(
-            report.descriptors[0].compatibility_mode,
-            PluginCompatibilityMode::OpenClawLegacy
-        );
-        assert_eq!(report.descriptors[0].manifest.plugin_id, "search");
-        assert_eq!(
-            report.descriptors[0].package_manifest_path,
-            Some(package_json.display().to_string())
-        );
-        let foreign = scan_diagnostic(
-            &report,
-            PluginDiagnosticCode::ForeignDialectContract,
-            "search",
-        )
-        .expect("foreign dialect diagnostic");
-        assert_eq!(foreign.phase, PluginDiagnosticPhase::Scan);
-        let legacy = scan_diagnostic(
-            &report,
-            PluginDiagnosticCode::LegacyOpenClawContract,
-            "search",
-        )
-        .expect("legacy openclaw diagnostic");
-        assert_eq!(legacy.phase, PluginDiagnosticPhase::Scan);
-        assert_eq!(
-            report.descriptors[0]
-                .manifest
-                .metadata
-                .get("adapter_family")
-                .map(String::as_str),
-            Some(OPENCLAW_LEGACY_COMPATIBILITY_ADAPTER_FAMILY)
-        );
-    }
-
-    #[test]
-    fn scanner_absorbs_plugins_into_catalog_and_pack() {
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![PluginDescriptor {
-                path: "/tmp/openai.rs".to_owned(),
-                source_kind: PluginSourceKind::EmbeddedSource,
-                dialect: PluginContractDialect::LoongEmbeddedSource,
-                dialect_version: None,
-                compatibility_mode: PluginCompatibilityMode::Native,
-                package_root: "/tmp".to_owned(),
-                package_manifest_path: None,
-                language: "rs".to_owned(),
-                manifest: PluginManifest {
-                    api_version: None,
-                    version: Some("1.3.0".to_owned()),
-                    plugin_id: "openai-rs".to_owned(),
-                    provider_id: "openai".to_owned(),
-                    connector_name: "openai".to_owned(),
-                    channel_id: Some("chat-main".to_owned()),
-                    endpoint: Some("https://api.openai.com/v1/chat/completions".to_owned()),
-                    capabilities: BTreeSet::from([
-                        Capability::InvokeConnector,
-                        Capability::ObserveTelemetry,
-                    ]),
-                    trust_tier: PluginTrustTier::Official,
-                    metadata: BTreeMap::from([("version".to_owned(), "1.3.0".to_owned())]),
-                    summary: None,
-                    tags: Vec::new(),
-                    input_examples: Vec::new(),
-                    output_examples: Vec::new(),
-                    defer_loading: false,
-                    setup: None,
-                    slot_claims: Vec::new(),
-                    compatibility: None,
-                },
-            }],
-        };
-
-        let mut catalog = IntegrationCatalog::new();
-        let mut pack = sample_pack();
-        let scanner = PluginScanner::new();
-
-        let absorb = scanner
-            .absorb(&mut catalog, &mut pack, &report)
-            .expect("absorb should succeed");
-        assert_eq!(absorb.absorbed_plugins, 1);
-        assert_eq!(absorb.provider_upserts, 1);
-        assert_eq!(absorb.channel_upserts, 1);
-        assert!(catalog.provider("openai").is_some());
-        assert!(catalog.channel("chat-main").is_some());
-        assert!(pack.allowed_connectors.contains("openai"));
-        assert!(
-            pack.granted_capabilities
-                .contains(&Capability::InvokeConnector)
-        );
-    }
-
-    #[test]
-    fn absorb_rejects_conflicting_exclusive_slot_claims() {
-        let report = PluginScanReport {
-            scanned_files: 2,
-            matched_plugins: 2,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![
-                PluginDescriptor {
-                    path: "/tmp/search-a.py".to_owned(),
-                    source_kind: PluginSourceKind::EmbeddedSource,
-                    dialect: PluginContractDialect::LoongEmbeddedSource,
-                    dialect_version: None,
-                    compatibility_mode: PluginCompatibilityMode::Native,
-                    package_root: "/tmp".to_owned(),
-                    package_manifest_path: None,
-                    language: "py".to_owned(),
-                    manifest: PluginManifest {
-                        api_version: None,
-                        version: None,
-                        plugin_id: "search-a".to_owned(),
-                        provider_id: "search-a".to_owned(),
-                        connector_name: "search-a".to_owned(),
-                        channel_id: None,
-                        endpoint: None,
-                        capabilities: BTreeSet::from([Capability::InvokeConnector]),
-                        trust_tier: PluginTrustTier::Unverified,
-                        metadata: BTreeMap::new(),
-                        summary: None,
-                        tags: Vec::new(),
-                        input_examples: Vec::new(),
-                        output_examples: Vec::new(),
-                        defer_loading: false,
-                        setup: None,
-                        slot_claims: vec![PluginSlotClaim {
-                            slot: "provider:web_search".to_owned(),
-                            key: "tavily".to_owned(),
-                            mode: PluginSlotMode::Exclusive,
-                        }],
-                        compatibility: None,
-                    },
-                },
-                PluginDescriptor {
-                    path: "/tmp/search-b.py".to_owned(),
-                    source_kind: PluginSourceKind::EmbeddedSource,
-                    dialect: PluginContractDialect::LoongEmbeddedSource,
-                    dialect_version: None,
-                    compatibility_mode: PluginCompatibilityMode::Native,
-                    package_root: "/tmp".to_owned(),
-                    package_manifest_path: None,
-                    language: "py".to_owned(),
-                    manifest: PluginManifest {
-                        api_version: None,
-                        version: None,
-                        plugin_id: "search-b".to_owned(),
-                        provider_id: "search-b".to_owned(),
-                        connector_name: "search-b".to_owned(),
-                        channel_id: None,
-                        endpoint: None,
-                        capabilities: BTreeSet::from([Capability::InvokeConnector]),
-                        trust_tier: PluginTrustTier::Unverified,
-                        metadata: BTreeMap::new(),
-                        summary: None,
-                        tags: Vec::new(),
-                        input_examples: Vec::new(),
-                        output_examples: Vec::new(),
-                        defer_loading: false,
-                        setup: None,
-                        slot_claims: vec![PluginSlotClaim {
-                            slot: "provider:web_search".to_owned(),
-                            key: "tavily".to_owned(),
-                            mode: PluginSlotMode::Exclusive,
-                        }],
-                        compatibility: None,
-                    },
-                },
-            ],
-        };
-
-        let mut catalog = IntegrationCatalog::new();
-        let mut pack = sample_pack();
-
-        let error = PluginScanner::new()
-            .absorb(&mut catalog, &mut pack, &report)
-            .expect_err("conflicting exclusive slot claims should fail");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("slot claim conflict"));
-        assert!(catalog.provider("search-a").is_none());
-        assert!(catalog.provider("search-b").is_none());
-    }
-
-    #[test]
-    fn absorb_allows_shared_and_advisory_slot_claims_and_projects_metadata() {
-        let current_host_version_req = format!(">={}", env!("CARGO_PKG_VERSION"));
-        let report = PluginScanReport {
-            scanned_files: 2,
-            matched_plugins: 2,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![
-                PluginDescriptor {
-                    path: "/tmp/search-shared.py".to_owned(),
-                    source_kind: PluginSourceKind::EmbeddedSource,
-                    dialect: PluginContractDialect::LoongEmbeddedSource,
-                    dialect_version: None,
-                    compatibility_mode: PluginCompatibilityMode::Native,
-                    package_root: "/tmp".to_owned(),
-                    package_manifest_path: None,
-                    language: "py".to_owned(),
-                    manifest: PluginManifest {
-                        api_version: None,
-                        version: Some("1.0.0".to_owned()),
-                        plugin_id: "search-shared".to_owned(),
-                        provider_id: "search-shared".to_owned(),
-                        connector_name: "search-shared".to_owned(),
-                        channel_id: None,
-                        endpoint: None,
-                        capabilities: BTreeSet::from([Capability::InvokeConnector]),
-                        trust_tier: PluginTrustTier::Unverified,
-                        metadata: BTreeMap::new(),
-                        summary: None,
-                        tags: Vec::new(),
-                        input_examples: Vec::new(),
-                        output_examples: Vec::new(),
-                        defer_loading: false,
-                        setup: None,
-                        slot_claims: vec![PluginSlotClaim {
-                            slot: "tool:search".to_owned(),
-                            key: "web".to_owned(),
-                            mode: PluginSlotMode::Shared,
-                        }],
-                        compatibility: Some(PluginCompatibility {
-                            host_api: Some(CURRENT_PLUGIN_HOST_API.to_owned()),
-                            host_version_req: Some(current_host_version_req.clone()),
-                        }),
-                    },
-                },
-                PluginDescriptor {
-                    path: "/tmp/search-advisory.py".to_owned(),
-                    source_kind: PluginSourceKind::EmbeddedSource,
-                    dialect: PluginContractDialect::LoongEmbeddedSource,
-                    dialect_version: None,
-                    compatibility_mode: PluginCompatibilityMode::Native,
-                    package_root: "/tmp".to_owned(),
-                    package_manifest_path: None,
-                    language: "py".to_owned(),
-                    manifest: PluginManifest {
-                        api_version: None,
-                        version: None,
-                        plugin_id: "search-advisory".to_owned(),
-                        provider_id: "search-advisory".to_owned(),
-                        connector_name: "search-advisory".to_owned(),
-                        channel_id: None,
-                        endpoint: None,
-                        capabilities: BTreeSet::from([Capability::InvokeConnector]),
-                        trust_tier: PluginTrustTier::Unverified,
-                        metadata: BTreeMap::new(),
-                        summary: None,
-                        tags: Vec::new(),
-                        input_examples: Vec::new(),
-                        output_examples: Vec::new(),
-                        defer_loading: false,
-                        setup: None,
-                        slot_claims: vec![PluginSlotClaim {
-                            slot: "tool:search".to_owned(),
-                            key: "web".to_owned(),
-                            mode: PluginSlotMode::Advisory,
-                        }],
-                        compatibility: None,
-                    },
-                },
-            ],
-        };
-
-        let mut catalog = IntegrationCatalog::new();
-        let mut pack = sample_pack();
-
-        let absorb = PluginScanner::new()
-            .absorb(&mut catalog, &mut pack, &report)
-            .expect("shared and advisory slot claims should coexist");
-
-        assert_eq!(absorb.absorbed_plugins, 2);
-        let shared_provider = catalog
-            .provider("search-shared")
-            .expect("shared provider should be registered");
-        assert_eq!(
-            shared_provider
-                .metadata
-                .get(PLUGIN_SLOT_CLAIMS_METADATA_KEY)
-                .map(String::as_str),
-            Some("[{\"slot\":\"tool:search\",\"key\":\"web\",\"mode\":\"shared\"}]")
-        );
-        assert_eq!(
-            shared_provider
-                .metadata
-                .get(PLUGIN_COMPATIBILITY_HOST_API_METADATA_KEY)
-                .map(String::as_str),
-            Some(CURRENT_PLUGIN_HOST_API)
-        );
-        assert_eq!(
-            shared_provider
-                .metadata
-                .get(PLUGIN_COMPATIBILITY_HOST_VERSION_REQ_METADATA_KEY)
-                .map(String::as_str),
-            Some(current_host_version_req.as_str())
-        );
-    }
-
-    #[test]
-    fn absorb_rejects_incompatible_host_api() {
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![PluginDescriptor {
-                path: "/tmp/incompatible-host.py".to_owned(),
-                source_kind: PluginSourceKind::EmbeddedSource,
-                dialect: PluginContractDialect::LoongEmbeddedSource,
-                dialect_version: None,
-                compatibility_mode: PluginCompatibilityMode::Native,
-                package_root: "/tmp".to_owned(),
-                package_manifest_path: None,
-                language: "py".to_owned(),
-                manifest: PluginManifest {
-                    api_version: None,
-                    version: None,
-                    plugin_id: "incompatible-host".to_owned(),
-                    provider_id: "incompatible-host".to_owned(),
-                    connector_name: "incompatible-host".to_owned(),
-                    channel_id: None,
-                    endpoint: None,
-                    capabilities: BTreeSet::from([Capability::InvokeConnector]),
-                    trust_tier: PluginTrustTier::Unverified,
-                    metadata: BTreeMap::new(),
-                    summary: None,
-                    tags: Vec::new(),
-                    input_examples: Vec::new(),
-                    output_examples: Vec::new(),
-                    defer_loading: false,
-                    setup: None,
-                    slot_claims: Vec::new(),
-                    compatibility: Some(PluginCompatibility {
-                        host_api: Some("loong-plugin/v999".to_owned()),
-                        host_version_req: None,
-                    }),
-                },
-            }],
-        };
-
-        let mut catalog = IntegrationCatalog::new();
-        let mut pack = sample_pack();
-
-        let error = PluginScanner::new()
-            .absorb(&mut catalog, &mut pack, &report)
-            .expect_err("incompatible host api should fail closed");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("compatibility.host_api"));
-        assert!(rendered.contains(CURRENT_PLUGIN_HOST_API));
-        assert!(catalog.provider("incompatible-host").is_none());
-    }
-
-    #[test]
-    fn absorb_rejects_invalid_host_version_requirement() {
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![PluginDescriptor {
-                path: "/tmp/invalid-version.py".to_owned(),
-                source_kind: PluginSourceKind::EmbeddedSource,
-                dialect: PluginContractDialect::LoongEmbeddedSource,
-                dialect_version: None,
-                compatibility_mode: PluginCompatibilityMode::Native,
-                package_root: "/tmp".to_owned(),
-                package_manifest_path: None,
-                language: "py".to_owned(),
-                manifest: PluginManifest {
-                    api_version: None,
-                    version: None,
-                    plugin_id: "invalid-version".to_owned(),
-                    provider_id: "invalid-version".to_owned(),
-                    connector_name: "invalid-version".to_owned(),
-                    channel_id: None,
-                    endpoint: None,
-                    capabilities: BTreeSet::from([Capability::InvokeConnector]),
-                    trust_tier: PluginTrustTier::Unverified,
-                    metadata: BTreeMap::new(),
-                    summary: None,
-                    tags: Vec::new(),
-                    input_examples: Vec::new(),
-                    output_examples: Vec::new(),
-                    defer_loading: false,
-                    setup: None,
-                    slot_claims: Vec::new(),
-                    compatibility: Some(PluginCompatibility {
-                        host_api: Some(CURRENT_PLUGIN_HOST_API.to_owned()),
-                        host_version_req: Some("not-a-semver-req".to_owned()),
-                    }),
-                },
-            }],
-        };
-
-        let mut catalog = IntegrationCatalog::new();
-        let mut pack = sample_pack();
-
-        let error = PluginScanner::new()
-            .absorb(&mut catalog, &mut pack, &report)
-            .expect_err("invalid host version requirement should fail closed");
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("compatibility.host_version_req"));
-        assert!(rendered.contains("invalid"));
-        assert!(catalog.provider("invalid-version").is_none());
-    }
-
-    #[test]
-    fn scanner_skips_non_utf8_files_instead_of_failing() {
-        let root = unique_tmp_dir("loong-plugin-binary");
-        fs::create_dir_all(&root).expect("create temp root");
-        let binary = root.join("compiled.bin");
-        fs::write(&binary, [0xff_u8, 0xfe, 0x00, 0x81]).expect("write binary file");
-
-        let scanner = PluginScanner::new();
-        let report = scanner
-            .scan_path(&root)
-            .expect("binary files should be skipped, not fail");
-        assert_eq!(report.scanned_files, 1);
-        assert_eq!(report.matched_plugins, 0);
-    }
-
-    #[test]
-    fn absorb_rolls_back_catalog_and_pack_on_validation_failure() {
-        // First descriptor is valid, second has an empty provider_id which
-        // triggers validation failure. The rollback must undo the first
-        // descriptor's mutations so catalog and pack remain unchanged.
-        let report = PluginScanReport {
-            scanned_files: 2,
-            matched_plugins: 2,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![
-                PluginDescriptor {
-                    path: "/tmp/good.rs".to_owned(),
-                    source_kind: PluginSourceKind::EmbeddedSource,
-                    dialect: PluginContractDialect::LoongEmbeddedSource,
-                    dialect_version: None,
-                    compatibility_mode: PluginCompatibilityMode::Native,
-                    package_root: "/tmp".to_owned(),
-                    package_manifest_path: None,
-                    language: "rs".to_owned(),
-                    manifest: PluginManifest {
-                        api_version: None,
-                        version: Some("1.0.0".to_owned()),
-                        plugin_id: "good-plugin".to_owned(),
-                        provider_id: "good-provider".to_owned(),
-                        connector_name: "good-connector".to_owned(),
-                        channel_id: Some("good-channel".to_owned()),
-                        endpoint: Some("https://good.local/invoke".to_owned()),
-                        capabilities: BTreeSet::from([Capability::InvokeConnector]),
-                        trust_tier: PluginTrustTier::VerifiedCommunity,
-                        metadata: BTreeMap::from([("version".to_owned(), "1.0.0".to_owned())]),
-                        summary: None,
-                        tags: Vec::new(),
-                        input_examples: Vec::new(),
-                        output_examples: Vec::new(),
-                        defer_loading: false,
-                        setup: None,
-                        slot_claims: Vec::new(),
-                        compatibility: None,
-                    },
-                },
-                PluginDescriptor {
-                    path: "/tmp/bad.rs".to_owned(),
-                    source_kind: PluginSourceKind::EmbeddedSource,
-                    dialect: PluginContractDialect::LoongEmbeddedSource,
-                    dialect_version: None,
-                    compatibility_mode: PluginCompatibilityMode::Native,
-                    package_root: "/tmp".to_owned(),
-                    package_manifest_path: None,
-                    language: "rs".to_owned(),
-                    manifest: PluginManifest {
-                        api_version: None,
-                        version: None,
-                        plugin_id: "bad-plugin".to_owned(),
-                        provider_id: String::new(), // empty — triggers validation error
-                        connector_name: "bad-connector".to_owned(),
-                        channel_id: None,
-                        endpoint: None,
-                        capabilities: BTreeSet::new(),
-                        trust_tier: PluginTrustTier::Unverified,
-                        metadata: BTreeMap::new(),
-                        summary: None,
-                        tags: Vec::new(),
-                        input_examples: Vec::new(),
-                        output_examples: Vec::new(),
-                        defer_loading: false,
-                        setup: None,
-                        slot_claims: Vec::new(),
-                        compatibility: None,
-                    },
-                },
-            ],
-        };
-
-        let mut catalog = IntegrationCatalog::new();
-        let mut pack = sample_pack();
-        let scanner = PluginScanner::new();
-
-        let catalog_before = catalog.clone();
-        let pack_before = pack.clone();
-
-        let result = scanner.absorb(&mut catalog, &mut pack, &report);
-        assert!(result.is_err(), "absorb should fail on empty provider_id");
-
-        // Verify rollback: catalog and pack are identical to their pre-absorb state.
-        assert_eq!(catalog, catalog_before, "catalog must be rolled back");
-        assert_eq!(pack, pack_before, "pack must be rolled back");
-    }
-
-    #[test]
-    fn format_plugin_provenance_summary_prefers_package_manifest_context() {
-        let summary = format_plugin_provenance_summary(
-            PluginSourceKind::EmbeddedSource,
-            "/tmp/pkg/plugin.py",
-            Some("/tmp/pkg/loong.plugin.json"),
-        );
-
-        assert_eq!(
-            summary,
-            "embedded_source:/tmp/pkg/plugin.py (package_manifest:/tmp/pkg/loong.plugin.json)"
-        );
-    }
-}
+mod tests;
